+++ client/src/api/dashboard.js
@@ -1,4 +1,4 @@
-import { get, post } from './base'
+import { get, patch, post } from './base'
 
 export async function getUpdates (pagenumber) {
   return (await get(`/../xhrapp.php?app=activity&m=load&page=${pagenumber}`)).data.updates
@@ -9,20 +9,19 @@ export async function getOptionListings () {
 }
 
 export async function saveOptionListings (options) {
-  var optionsString = ''
-  var newOptionIdInx = 0
-
+  const request = []
   for (var optionId in options) {
     options[optionId].items = options[optionId].items.filter((a) => { return !a.checked })
     for (var item in options[optionId].items) {
-      optionsString += '&options[' + newOptionIdInx + '][index]=' + options[optionId].index + '&options[' + newOptionIdInx + '][id]=' + options[optionId].items[item].id
-      newOptionIdInx += 1
+      request.push({
+        index: options[optionId].index,
+        id: options[optionId].items[item].id
+      })
     }
   }
-  if (optionsString === '') {
-    optionsString = '&select_all_options=true'
-  }
-  return get(`/../xhrapp.php?app=activity&m=setoptionlist${optionsString}`)
+  return patch('/activities/options', {
+    options: request
+  })
 }
 
 export async function sendQuickreply (href, msg) {